add fragment to MainActivity
