+++ app/src/main/java/com/brahmachilakala/bakingapp/MainActivity.java
@@ -1,121 +1,27 @@
 package com.brahmachilakala.bakingapp;
 
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
+import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.GestureDetector;
-import android.view.MotionEvent;
-import android.view.View;
 
-import org.json.JSONArray;
-
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.ArrayList;
-
-public class MainActivity extends AppCompatActivity {
-
-    private RecyclerView rvRecipes;
-    private RecipesAdapter mRecipesAdapter;
-    ArrayList<Recipe> mRecipes;
-    private GestureDetector mGestureDetector;
+public class MainActivity extends AppCompatActivity implements MainFragment.SendIntentListener{
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        new RecipeTask().execute("https://d17h27t6h515a5.cloudfront.net/topher/2017/May/59121517_baking/baking.json");
+        FragmentTransaction ft = getSupportFragmentManager().beginTransaction();
+        ft.replace(R.id.fragment_main, new MainFragment())
+                .commit();
 
     }
 
-    public void loadRecyclerView() {
-        rvRecipes = (RecyclerView) findViewById(R.id.rvRecipes);
-        mRecipesAdapter = new RecipesAdapter(mRecipes);
-
-        rvRecipes.setAdapter(mRecipesAdapter);
-        rvRecipes.setLayoutManager(new LinearLayoutManager(this));
-
-        mGestureDetector = new GestureDetector(this, new GestureDetector.SimpleOnGestureListener() {
-            @Override
-            public boolean onSingleTapUp(MotionEvent e) {
-                return true;
-            }
-        });
-
-        rvRecipes.addOnItemTouchListener(new RecyclerView.OnItemTouchListener() {
-            @Override
-            public boolean onInterceptTouchEvent(RecyclerView rv, MotionEvent e) {
-                View childView = rv.findChildViewUnder(e.getX(), e.getY());
-
-                if (childView != null && mGestureDetector.onTouchEvent(e)) {
-                    int position = rv.getChildAdapterPosition(childView);
-
-                    Intent recipeDetailActivity = new Intent(MainActivity.this, RecipeDetailActivity.class);
-                    recipeDetailActivity.putExtra("recipe", mRecipes.get(position));
-
-                    startActivity(recipeDetailActivity);
-
-                    return true;
-                }
-                return false;
-            }
-
-            @Override
-            public void onTouchEvent(RecyclerView rv, MotionEvent e) {
-
-            }
-
-            @Override
-            public void onRequestDisallowInterceptTouchEvent(boolean disallowIntercept) {
-
-            }
-        });
-    }
-
-    private class RecipeTask extends AsyncTask<String, Void, String> {
-
-        @Override
-        protected String doInBackground(String... strings) {
-            try {
-                URL url = new URL(strings[0]);
-                HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-                conn.connect();
-                InputStream in = conn.getInputStream();
-                BufferedReader reader = new BufferedReader(new InputStreamReader(in));
-
-                StringBuilder stringBuilder = new StringBuilder();
-
-                String line = "";
-                while ((line = reader.readLine()) != null) {
-                    stringBuilder.append(line);
-                }
-
-                return stringBuilder.toString();
-
-
-            } catch (Exception e) {
-                Log.i("MainActivity", "Error in fetching contents : " + e.getMessage());
-                return null;
-            }
-        }
-
-        @Override
-        protected void onPostExecute(String s) {
-            try {
-                JSONArray jsonArray = new JSONArray(s);
-                mRecipes = Recipe.fromRecipeJson(jsonArray);
-                loadRecyclerView();
-            } catch (Exception e) {
-                Log.i("MainActivity", "Error in parsing JSON in Recipe Task" + e.getMessage());
-            }
-        }
+    @Override
+    public void startIntent(Recipe recipe) {
+        Intent recipeDetailActivity = new Intent(MainActivity.this, RecipeDetailActivity.class);
+        recipeDetailActivity.putExtra("recipe", recipe);
+        startActivity(recipeDetailActivity);
     }
 }
